+++ src/models/db.js
@@ -4,11 +4,13 @@ import { categoryMemStore } from "./mem/category-mem-store.js";
 import { poiMemStore } from "./mem/poi-mem-store.js";
 
 import { userJsonStore } from "./json/user-json-store.js";
-import { playlistJsonStore } from "./json/playlist-json-store.js";
-import { trackJsonStore } from "./json/track-json-store.js";
+import { categoryJsonStore } from "./json/category-json-store.js";
+import { poiJsonStore } from "./json/poi-json-store.js";
+
+
 import { userMongoStore } from "./mongo/user-mongo-store.js";
-import { playlistMongoStore } from "./mongo/playlist-mongo-store.js";
-import { trackMongoStore } from "./mongo/track-mongo-store.js";
+import { categoryMongoStore } from "./mongo/category-mongo-store.js";
+import { poiMongoStore } from "./mongo/poi-mongo-store.js";
 import { connectMongo } from "./mongo/connect.js";
 
 export const db = {
@@ -20,13 +22,13 @@ export const db = {
     switch (storeType) {
       case "json" :
         this.userStore = userJsonStore;
-        this.playlistStore = playlistJsonStore;
-        this.poiStore = trackJsonStore;
+        this.categoryStore = categoryJsonStore;
+        this.poiStore = poiJsonStore;
         break;
       case "mongo" :
         this.userStore = userMongoStore;
-        this.playlistStore = playlistMongoStore;
-        this.poiStore = trackMongoStore;
+        this.categoryStore = categoryMongoStore;
+        this.poiStore = poiMongoStore;
         connectMongo();
         break;
       default :