+++ test/models/poi-model-test.js
@@ -1,19 +1,18 @@
 import { assert } from "chai";
 import { db } from "../../src/models/db.js";
-import { testCategories, testPOI, Kiln,testKiln } from "../fixtures.js";
+import { testDBType, testCategories, testPOI, Kiln, testKiln } from "../fixtures.js";
 import { assertSubset } from "../test-utils.js";
 
 suite("POI Model tests", () => {
-
   let categoryList = null;
 
   setup(async () => {
-    db.init("mem");
+    db.init(testDBType);
     // delete all catagories form the list
     await db.categoryStore.deleteAllCategories();
     await db.poiStore.deleteAllPOI();
     // add a new category
-    categoryList = await db.categoryStore.addCategory(testCategories);
+    categoryList = await db.categoryStore.addCategory(Kiln);
     // add the test points to the category
     for (let i = 0; i < testPOI.length; i += 1) {
       // eslint-disable-next-line no-await-in-loop
@@ -21,31 +20,30 @@ suite("POI Model tests", () => {
     }
   });
 
-  test("create single POI", async () => {
+  test("Create single POI", async () => {
     const kilnList = await db.categoryStore.addCategory(Kiln);
-    const poi = await db.poiStore.addPOI(kilnList._id, testKiln)
+    const poi = await db.poiStore.addPOI(kilnList._id, testKiln);
     assert.isNotNull(poi._id);
-    assertSubset (testKiln, poi);
+    assertSubset(testKiln, poi);
   });
 
   test("Read multiple POIs", async () => {
-    const aPOI = await db.categoryStore.getCategoryById(categoryList._id);
-    assert.equal(testPOI.length, aPOI.length);
+    assert.equal(1, 6);
   });
 
   test("delete all POIs", async () => {
     const aPOIs = await db.poiStore.getAllPOI();
     assert.equal(testPOI.length, aPOIs.length);
     await db.poiStore.deleteAllPOI();
-    const newPOIS = await db.poiStore.getAllPOI();
-    assert.equal(0, newPOIS.length);
+    const newPOI = await db.poiStore.getAllPOI();
+    assert.equal(0, newPOI.length);
   });
 
   test("get a POI - success", async () => {
     const testCategory = await db.categoryStore.addCategory(Kiln);
-    const poi = await db.poiStore.addPOI(testCategory._id, testKiln)
+    const poi = await db.poiStore.addPOI(testCategory._id, testKiln);
     const newPOI = await db.poiStore.getPOIById(poi._id);
-    assertSubset (testKiln, newPOI);
+    assertSubset(testKiln, newPOI);
   });
 
   test("delete One POI - success", async () => {