+++ test/models/users-model-test.js
@@ -1,12 +1,12 @@
 import { assert } from "chai";
 import { db } from "../../src/models/db.js";
-import { maggie, testUsers } from "../fixtures.js";
+import { testDBType,maggie, testUsers } from "../fixtures.js";
 import { assertSubset } from "../test-utils.js";
 
 suite("User Model tests", () => {
 
   setup(async () => {
-    db.init("mongo");
+    db.init(testDBType);
     await db.userStore.deleteAll();
     for (let i = 0; i < testUsers.length; i += 1) {
       // eslint-disable-next-line no-await-in-loop